+++ tests/data/serialization/labelbox_v1/test_image.py
@@ -9,6 +9,10 @@ from labelbox.data.serialization.labelbox_v1.converter import LBV1Converter
     'tests/data/assets/labelbox_v1/highly_nested_image.json',
     'tests/data/assets/labelbox_v1/image_export.json'
 ])
+#TODO: some checklists from the export come in as [checklist ans: []]
+# while others are checklist ans: []... when we can figure out why we sometimes
+# have extra brackets, we can look into testing nested checklist answers
+# and ensuring the export's output matches deserialized/serialized output
 def test_image(file_path):
     with open(file_path, 'r') as file:
         payload = json.load(file)
@@ -16,7 +20,11 @@ def test_image(file_path):
     collection = LBV1Converter.deserialize([payload])
     serialized = next(LBV1Converter.serialize(collection))
 
+    # We are storing the media types now.
+    payload['media_type'] = 'image'
+
     assert serialized.keys() == payload.keys()
+
     for key in serialized:
         if key != 'Label':
             assert serialized[key] == payload[key]
@@ -32,7 +40,6 @@ def test_image(file_path):
                     if isinstance(annotation_b['classifications'][0], list):
                         annotation_b['classifications'] = annotation_b[
                             'classifications'][0]
-
                 assert annotation_a == annotation_b