+++ tests/integration/test_ontology.py
@@ -2,6 +2,9 @@ import pytest
 
 from labelbox.exceptions import InconsistentOntologyException
 from labelbox import Tool, Classification, Option, OntologyBuilder
+from labelbox.orm.model import Entity
+import json
+import time
 
 _SAMPLE_ONTOLOGY = {
     "tools": [{
@@ -122,13 +125,13 @@ _SAMPLE_ONTOLOGY = {
             "schemaNodeId": None,
             "featureSchemaId": None,
             "label": "yes",
-            "value": "yes",
+            "value": "definitely yes",
             "options": []
         }, {
             "schemaNodeId": None,
             "featureSchemaId": None,
             "label": "no",
-            "value": "no",
+            "value": "definitely not",
             "options": []
         }]
     }]
@@ -149,10 +152,20 @@ def test_create_classification(class_type) -> None:
 
 @pytest.mark.parametrize("value, expected_value, typing",
                          [(3, 3, int), ("string", "string", str)])
-def test_create_option(value, expected_value, typing) -> None:
+def test_create_option_with_value(value, expected_value, typing) -> None:
     o = Option(value=value)
     assert (o.value == expected_value)
     assert (o.value == o.label)
+
+
+@pytest.mark.parametrize("value, label, expected_value, typing",
+                         [(3, 2, 3, int),
+                          ("string", "another string", "string", str)])
+def test_create_option_with_value_and_label(value, label, expected_value,
+                                            typing) -> None:
+    o = Option(value=value, label=label)
+    assert (o.value == expected_value)
+    assert o.value != o.label
 
 
 def test_create_empty_ontology() -> None:
@@ -240,3 +253,74 @@ def test_ontology_asdict(project) -> None:
 def test_from_project_ontology(client, project) -> None:
     o = OntologyBuilder.from_project(project)
     assert o.asdict() == project.ontology().normalized
+
+
+def _get_attr_stringify_json(obj, attr):
+    value = getattr(obj, attr.name)
+    if attr.field_type.name.lower() == "json":
+        return json.dumps(value, sort_keys=True)
+    return value
+
+
+def test_feature_schema_create_read(client, rand_gen):
+    name = f"test-root-schema-{rand_gen(str)}"
+    feature_schema_cat_normalized = {
+        'tool': 'polygon',
+        'name': name,
+        'color': 'black',
+        'classifications': [],
+    }
+    created_feature_schema = client.create_feature_schema(
+        feature_schema_cat_normalized)
+    queried_feature_schema = client.get_feature_schema(
+        created_feature_schema.uid)
+    for attr in Entity.FeatureSchema.fields():
+        assert _get_attr_stringify_json(created_feature_schema,
+                                        attr) == _get_attr_stringify_json(
+                                            queried_feature_schema, attr)
+
+    time.sleep(3)  # Slight delay for searching
+    queried_feature_schemas = list(client.get_feature_schemas(name))
+    assert [feature_schema.name for feature_schema in queried_feature_schemas
+           ] == [name]
+    queried_feature_schema = queried_feature_schemas[0]
+
+    for attr in Entity.FeatureSchema.fields():
+        assert _get_attr_stringify_json(created_feature_schema,
+                                        attr) == _get_attr_stringify_json(
+                                            queried_feature_schema, attr)
+
+
+def test_ontology_create_read(client, rand_gen):
+    ontology_name = f"test-ontology-{rand_gen(str)}"
+    tool_name = f"test-ontology-tool-{rand_gen(str)}"
+    feature_schema_cat_normalized = {
+        'tool': 'polygon',
+        'name': tool_name,
+        'color': 'black',
+        'classifications': [],
+    }
+    feature_schema = client.create_feature_schema(feature_schema_cat_normalized)
+    created_ontology = client.create_ontology_from_feature_schemas(
+        name=ontology_name, feature_schema_ids=[feature_schema.uid])
+    tool_normalized = created_ontology.normalized['tools'][0]
+    for k, v in feature_schema_cat_normalized.items():
+        assert tool_normalized[k] == v
+    assert tool_normalized['schemaNodeId'] is not None
+    assert tool_normalized['featureSchemaId'] == feature_schema.uid
+
+    queried_ontology = client.get_ontology(created_ontology.uid)
+
+    for attr in Entity.Ontology.fields():
+        assert _get_attr_stringify_json(created_ontology,
+                                        attr) == _get_attr_stringify_json(
+                                            queried_ontology, attr)
+
+    time.sleep(3)  # Slight delay for searching
+    queried_ontologies = list(client.get_ontologies(ontology_name))
+    assert [ontology.name for ontology in queried_ontologies] == [ontology_name]
+    queried_ontology = queried_ontologies[0]
+    for attr in Entity.Ontology.fields():
+        assert _get_attr_stringify_json(created_ontology,
+                                        attr) == _get_attr_stringify_json(
+                                            queried_ontology, attr)